requirements.txt
absl-py==1.4.0
asttokens==3.0.1
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.3
comtypes==1.1.14
contourpy==1.0.7
cvzone==1.5.6
cycler==0.11.0
debugpy==1.8.19
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.39.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==8.7.1
importlib-resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==2.15.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.2.1
mediapipe==0.10.11
ml-dtypes==0.2.0
nest-asyncio==1.6.0
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
parso==0.8.5
Pillow==9.4.0
platformdirs==4.4.0
prompt-toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.1
pure-eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.23
pyenchant==3.2.2
pygments==2.19.2
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pywin32==305
pyzmq==27.1.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.13.1
six==1.16.0
sounddevice==0.5.3
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tornado==6.5.4
traitlets==5.14.3
typing-extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.14
Werkzeug==2.2.3
wrapt==1.14.2
zipp==3.23.0
